requirements.txt
Flask
Flask-Session
psycopg2
SQLAlchemy

python-dotenv
flask-bcrypt
